% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorings.R
\name{a_crit}
\alias{a_crit}
\title{A-Criteria Score}
\usage{
a_crit(design, order)
}
\arguments{
\item{design}{the design matrix you want to score}

\item{order}{order of the model you want.
Enter 0 for a first order main effects model,
1 for a first order model with 2 way interactions,
2 for a second order model with 2 way interactions and squared main effects}
}
\value{
This function returns the trace of the inverse of the information
matrix.
}
\description{
This function takes in a design matrix and a desired model as input. From the
inputs, the design matrix is expanded into the model matrix and then the
information matrix is computed. From that, inverse of the information matrix
is computed and the trace of the inverse of the information matrix is
calculated.
}
