% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coordinate_D}
\alias{coordinate_D}
\title{Coordinate Exchange Algorithm}
\usage{
coordinate_D(N, K, model_order)
}
\arguments{
\item{N}{the number of trials in the experiment}

\item{K}{the number of factors in the experiment}

\item{model_order}{order of the model that you want.
Enter 0 for a first order main effects model, 1 for a first order model
with 2 way interactions,
2 for a second order model with 2 way interactions and
squared main effects}
}
\value{
a list of outputs that tells you the initial design and score,
  and then the optimized design and score
}
\description{
Coordinate Exchange Algorithm
}
