% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDS.R
\name{fds_plot}
\alias{fds_plot}
\title{Fraction of Design Space Plots}
\usage{
fds_plot(des_list, order)
}
\arguments{
\item{des_list}{a list of design matrices that you want to compare}

\item{order}{order of the model you want.
Enter 0 for a first order main effects model,
1 for a first order model with 2 way interactions,
2 for a second order model with 2 way interactions and squared main effects}
}
\value{
This function returns a plot of relative prediction variances over
the design space for given experimental design matrices.
}
\description{
Generate Fraction of Design Space plots for a list of design matrices
}
\examples{
# Generate Optimal Designs
optimal_designs <- generate_designs(n = 15, k = 3, order = 0, criteria = "A",
 iterations = 10)

# Pull out the 4 lowest A-Scored designs
result <- lapply(head(optimal_designs, 4), `[[`, 1)

# Plot those designs
fds_plot(result, order = 0)

}
