% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorings.R
\name{model_matrix}
\alias{model_matrix}
\title{Model Matrix}
\usage{
model_matrix(design, order)
}
\arguments{
\item{design}{the design matrix}

\item{order}{order of the model you want}
}
\description{
Model Matrix
}
