% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{optimal_A_14_3}
\alias{optimal_A_14_3}
\title{optimal_A_14_3}
\format{
## `optimal_A_14_3`
A list with 1000 entries, each entry is a list with a length of 2:
\describe{
  \item{entry 1}{Locally A-optimal design matrix}
  \item{entry 2}{The associated score with the design matrix}
  ...
}
}
\usage{
optimal_A_14_3
}
\description{
A list of generated A-optimal experimental designs.
}
\keyword{datasets}
