% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{optimal_D_25_4}
\alias{optimal_D_25_4}
\title{optimal_D_25_4}
\format{
## `optimal_D_25_4`
A list with 100 entries, each entry is a list with a length of 2:
\describe{
  \item{entry 1}{Locally D-optimal design matrix with 25 rows and 4 columns}
  \item{entry 2}{The associated score with the design matrix}
  ...
}
}
\usage{
optimal_D_25_4
}
\description{
A list of generated D-optimal experimental designs.
}
\keyword{datasets}
