% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorings.R
\name{singular_check}
\alias{singular_check}
\title{Check Singularity of an information matrix from a design matrix.}
\usage{
singular_check(X, order)
}
\arguments{
\item{X}{A design matrix with n rows and k columns.}

\item{order}{Model you want to fit. 0 fits a first order model. 1 fits a
first order model with two-way interactions. 2 fits a second order model.}
}
\value{
This function returns a TRUE if the information matrix created from
the design matrix is singular, FALSE if not.
}
\description{
This function takes in a design matrix and a model desired to be fit. From
this, we expand the design out into the model matrix and then compute the
information matrix. This function then checks if the matrix is invertible.
}
