% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDS.R
\name{FDS_plot}
\alias{FDS_plot}
\title{Fraction of Design Space Plots}
\usage{
FDS_plot(Des_list, order)
}
\arguments{
\item{Des_list}{a list of design matrices that you want to compare}

\item{order}{order of the model you want.
Enter 0 for a first order main effects model,
1 for a first order model with 2 way interactions,
2 for a second order model with 2 way interactions and squared main effects}
}
\value{
a plot of relative prediction variances over the design space
}
\description{
Generate Fraction of Design Space plots for the design matrices
}
