% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_designs.R
\name{generate_designs}
\alias{generate_designs}
\title{Generate Optimal Designs by Coordinate Exchange}
\usage{
generate_designs(N, k, order = 1, criteria = "D", iterations = 1000)
}
\description{
Generate Optimal Designs by Coordinate Exchange
}
